fix: long -> doubleに変更

+++ app/src/main/java/com/example/yosuke/calculator/model/entity/CalcEntity.kt
@@ -1,6 +1,6 @@
 package com.example.yosuke.calculator.model.entity
 
-data class CalcEntity(val number: Long, var operator: Operators) {
+data class CalcEntity(val number: Double, var operator: Operators) {
     override fun toString(): String {
         return "$number ${operator.toStr()}"
     }
