+++ app/src/main/java/com/example/yosuke/calculator/model/usecase/CalcUseCase.kt
@@ -3,5 +3,5 @@ package com.example.yosuke.calculator.model.usecase
 import com.example.yosuke.calculator.model.entity.Operators
 
 interface CalcUseCase {
-    fun calc(x: Long, operator: Operators, y: Long): Long
+    fun calc(x: Double, operator: Operators, y: Double): Double
 }
