+++ app/src/main/java/com/example/yosuke/calculator/model/usecase/CalcUseCaseImpl.kt
@@ -3,7 +3,7 @@ package com.example.yosuke.calculator.model.usecase
 import com.example.yosuke.calculator.model.entity.Operators
 
 class CalcUseCaseImpl : CalcUseCase {
-    override fun calc(x: Long, operator: Operators, y: Long): Long {
+    override fun calc(x: Double, operator: Operators, y: Double): Double {
         return when (operator) {
             Operators.PLUS -> x + y
             Operators.MINUS -> x - y
